backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "redis>=6.4.0",
    "uvicorn>=0.35.0",
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
    "email-validator>=2.0.0",
    "pydantic>=2.11.7",
]
